requirements.txt
django==3.2.8
djangorestframework==3.13.1
drf_yasg==1.20.0
coverage==6.3.2
dj-database-url==0.5.0
psycopg2-binary==2.8.6
whitenoise==5.3.0
gunicorn==19.8.1
injector==0.18.4
json_log_formatter==0.4.0
django-cors-headers==3.11.0
django-safedelete==1.1.2
